Rebase + feedback
