+++ test/Microsoft.AspNetCore.SignalR.Tests/EndToEndTests.cs
@@ -5,33 +5,10 @@ using System;
 using System.Net.WebSockets;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 
 namespace Microsoft.AspNetCore.SignalR.Tests
 {
-
-    public class Startup
-    {
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddSockets();
-            services.AddSignalR();
-            services.AddSingleton<EchoEndPoint>();
-        }
-
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
-        {
-            app.UseSockets(options => options.MapEndpoint<EchoEndPoint>("/echo"));
-            app.UseSignalR(routes =>
-            {
-            });
-        }
-    }
-
-
     [CollectionDefinition(Name)]
     public class EndToEndTestsCollection : ICollectionFixture<ServerFixture>
     {
@@ -50,23 +27,21 @@ namespace Microsoft.AspNetCore.SignalR.Tests
 
         [Fact]
         public async Task WebSocketsTest()
+        {
+            const string message = "Hello, World!";
+            using (var ws = new ClientWebSocket())
             {
-                const string message = "Hello, World!";
-                using (var ws = new ClientWebSocket())
-                {
-                    await ws.ConnectAsync(new Uri(_serverFixture.WebSocketsUrl +"/echo/ws"), CancellationToken.None);
-                    var bytes = System.Text.Encoding.UTF8.GetBytes(message);
-                    await ws.SendAsync(new ArraySegment<byte>(bytes), WebSocketMessageType.Binary, true, CancellationToken.None);
+                await ws.ConnectAsync(new Uri(_serverFixture.WebSocketsUrl + "/echo/ws"), CancellationToken.None);
+                var bytes = System.Text.Encoding.UTF8.GetBytes(message);
+                await ws.SendAsync(new ArraySegment<byte>(bytes), WebSocketMessageType.Binary, true, CancellationToken.None);
 
-                    var buffer = new ArraySegment<byte>(new byte[1024]);
-                    var result = await ws.ReceiveAsync(buffer, CancellationToken.None);
+                var buffer = new ArraySegment<byte>(new byte[1024]);
+                var result = await ws.ReceiveAsync(buffer, CancellationToken.None);
 
-                    Assert.Equal(message, System.Text.Encoding.UTF8.GetString(buffer.Array, 0, result.Count));
+                Assert.Equal(message, System.Text.Encoding.UTF8.GetString(buffer.Array, 0, result.Count));
 
-                    await ws.CloseAsync(WebSocketCloseStatus.Empty, "", CancellationToken.None);
-                }
+                await ws.CloseAsync(WebSocketCloseStatus.Empty, "", CancellationToken.None);
             }
-
+        }
     }
-
 }
